Document QA app/requirements.txt
pinecone-client
pypdf
openai
tiktoken
langchain
pandas 
numpy
streamlit
